// src/app/observation/model/observationStatusFilter.model.ts
export class ObservationStatusFilter {
    OrderdBy: string;
    Observationtypes: number[];
    StatusIds: number[];
    PageSize: number;
    PageIndex: number;
    ObservationFromDate: Date;
    ObservationToDate: Date;
    ObservationSubmissionFromDate: Date;
    ObservationSubmissionToDate: Date;
}