// src/app/observation/model/observationStatusReport.model.ts
export class ObservationStatusReport {
    ObservationRequestId: number;
    DisplayId: string;
    RaisedByValue: string;
    RaisedBy: number;
    RaisedFor: number;
    ObservationDate: Date;
    CreatedDate: Date;
    ClosureDate: Date;
    DeletionDate: Date;
    Status: string;
    AcknowledgmentDate: Date;
    SubmittedByUserName: string;
    OnBehalfUserName: string;
    NoOfActionsAdded: number;
}